# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: Al-Shalal Logistics API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /revenues:
    get:
      operationId: listRevenues
      tags: [revenues]
      parameters:
        - name: driverId
          in: query
          schema:
            type: integer
        - name: activeOnly
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of revenues
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Revenue"
    post:
      operationId: createRevenue
      tags: [revenues]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRevenueBody"
      responses:
        "201":
          description: Created revenue
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Revenue"
  /revenues/{id}/repay:
    patch:
      operationId: repayRevenue
      tags: [revenues]
      summary: Mark a deferred revenue as paid (clears deferred flag and amount)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated revenue
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Revenue"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /revenues/{id}:
    patch:
      operationId: updateRevenue
      tags: [revenues]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRevenueBody"
      responses:
        "200":
          description: Updated revenue
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Revenue"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteRevenue
      tags: [revenues]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── OWNER TRANSFERS ──────────────────────────────────
  /transfers:
    get:
      operationId: listTransfers
      tags: [transfers]
      parameters:
        - name: driverId
          in: query
          schema:
            type: integer
        - name: activeOnly
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of transfers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transfer"
    post:
      operationId: createTransfer
      tags: [transfers]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTransferBody"
      responses:
        "201":
          description: Created transfer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transfer"
  /transfers/{id}:
    patch:
      operationId: updateTransfer
      tags: [transfers]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTransferBody"
      responses:
        "200":
          description: Updated transfer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transfer"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Cannot edit a settled transfer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteTransfer
      tags: [transfers]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── CYCLE SUMMARY ─────────────────────────────────────
  /cycle-summary:
    get:
      operationId: getCycleSummary
      tags: [settlements]
      parameters:
        - name: driverId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Current cycle summary for a driver
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CycleSummary"

  # ── SETTLEMENTS ──────────────────────────────────────
  /settlements:
    get:
      operationId: listSettlements
      tags: [settlements]
      parameters:
        - name: driverId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of past settlements
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Settlement"
    post:
      operationId: createSettlement
      tags: [settlements]
      summary: Settle current cycle for a driver
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSettlementBody"
      responses:
        "201":
          description: Settlement created and cycle archived
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Settlement"

  # ── ANALYTICS ─────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    Revenue:
      type: object
      properties:
        id:
          type: integer
        driverId:
          type: integer
        amount:
          type: number
        clientName:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        receiptImageUrl:
          type: string
          nullable: true
        date:
          type: string
          format: date
        settlementId:
          type: integer
          nullable: true
        isDeferred:
          type: boolean
        deferredAmount:
          type: number
        isPaid:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, driverId, amount, date, createdAt]
    CreateRevenueBody:
      type: object
      properties:
        driverId:
          type: integer
        amount:
          type: number
        clientName:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        receiptImageUrl:
          type: string
          nullable: true
        date:
          type: string
          format: date
        isDeferred:
          type: boolean
      required: [driverId, amount, date]
    UpdateRevenueBody:
      type: object
      properties:
        amount:
          type: number
        clientName:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        date:
          type: string
          format: date
    Transfer:
      type: object
      properties:
        id:
          type: integer
        driverId:
          type: integer
        amount:
          type: number
        description:
          type: string
          nullable: true
        receiptImageUrl:
          type: string
          nullable: true
        date:
          type: string
          format: date
        settlementId:
          type: integer
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, driverId, amount, date, createdAt]
    CreateTransferBody:
      type: object
      properties:
        driverId:
          type: integer
        amount:
          type: number
        description:
          type: string
          nullable: true
        receiptImageUrl:
          type: string
          nullable: true
        date:
          type: string
          format: date
      required: [driverId, amount, date]
    UpdateTransferBody:
      type: object
      properties:
        amount:
          type: number
        description:
          type: string
          nullable: true
        date:
          type: string
          format: date
    CycleSummary:
      type: object
      properties:
        driverId:
          type: integer
        totalRevenue:
          type: number
        totalExpenses:
          type: number
        totalTransfers:
          type: number
        netProfit:
          type: number
        driverShare:
          type: number
        ownerPayout:
          type: number
        revenueCount:
          type: integer
        expenseCount:
          type: integer
        transferCount:
          type: integer
        postponedBalance:
          type: number
        deferredCount:
          type: integer
        grossRevenue:
          type: number
        cashRevenue:
          type: number
      required:
        - driverId
        - totalRevenue
        - totalExpenses
        - totalTransfers
        - netProfit
        - driverShare
        - ownerPayout
        - revenueCount
        - expenseCount
        - transferCount
        - postponedBalance
        - deferredCount
        - grossRevenue
        - cashRevenue
    Settlement:
      type: object
      properties:
        id:
          type: integer
        driverId:
          type: integer
        totalRevenue:
          type: number
        totalExpenses:
          type: number
        totalTransfers:
          type: number
        netProfit:
          type: number
        driverShare:
          type: number
        ownerPayout:
          type: number
        periodStart:
          type: string
          format: date
          nullable: true
        periodEnd:
          type: string
          format: date
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - driverId
        - totalRevenue
        - totalExpenses
        - totalTransfers
        - netProfit
        - driverShare
        - ownerPayout
        - createdAt
    CreateSettlementBody:
      type: object
      properties:
        driverId:
          type: integer
      required: [driverId]
...